Changes to Profile, ProfilesManagement and Webserver
Profile now has the PrivacyLevel in the constructor, and Email is also in the constructor for Profiles. Whenever a user is trying to signup, they now need to include those 2 parameters instead of just display name, username, and password. This includes all of the functions in the profiles that didn't include these previously, so they're automatically being saved to MongoDB and being retrieved from MongoDB when a profile obj is being created.